requirements.txt
psycopg2_binary
streamlit
pandas
pandas-datareader
Pyarrow
yfinance